requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
langchain>=0.2.3
langchain-openai>=0.1.0
openai>=1.59.8
weaviate-client>=3.25.2
sqlalchemy>=2.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=0.8.11
unstructured>=0.10.0
easyocr>=1.7.0
streamlit>=1.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
websockets>=12.0
asyncio-mqtt>=0.16.0
uvicorn>=0.24.0
fastapi>=0.104.0 